WS2P SSL

+++ core/src/main/java/juniter/core/model/dto/wot/requirements/ReqIdtyDTO.java
@@ -24,6 +24,6 @@ public class ReqIdtyDTO {
     List<IdtyCerts> pendingCerts;
     List<IdtyCerts> pendingMemberships;
     Integer membershipPendingExpiresIn ;
-    Integer membershipExpiresIn;
+    Long membershipExpiresIn;
 
 }
